#0210-4:完成订单新增功能

+++ ruoyi-txs/src/main/java/com/ruoyi/txs/mapper/OrderProcessDao.java
@@ -18,5 +18,5 @@ public interface OrderProcessDao {
 
     int update(TxsOrderProcess param);
 
-    int insert(TxsOrderProcess param);
+    Long insert(TxsOrderProcess param);
 }
